requirements.txt
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.4
black==24.2.0
flake8==7.0.0
gunicorn==21.2.0
pytest-cov==4.1.0